refactor(clock): move the wall-clock reporters to konfai.utils.clock
SweepClock lived in the patching engine and StartupClock in the runtime, which imported it lazily to break the cycle. Both are wall-clock accounting used by every workflow; they now share one module with no import cycle.
